// src/app/models/attendance/cls_TRTimeallw.ts
export class cls_TRTimeallw {
    constructor() {

    }

    company_code: string = "";
    plantimeallw_code: string = "";
    timeallw_no: string = "";
    timeallw_time: string = "0";
    timeallw_type: string = "A";
    timeallw_method: string = "H";
    timeallw_timein: string = "00:00";
    timeallw_timeout: string = "00:00";
    timeallw_normalday: string = "0.00";
    timeallw_offday: string = "0.00";
    timeallw_companyday: string = "0.00";
    timeallw_holiday: string = "0.00";
    timeallw_leaveday: string = "0.00";
}